P3A: Extract some code to a separate file.
